yul-phaser: Unsigned abs() of a difference of unsigned values

+++ test/yulPhaser/TestHelpers.cpp
@@ -74,7 +74,10 @@ size_t phaser::test::countDifferences(Chromosome const& _chromosome1, Chromosome
 	for (size_t i = 0; i < min(_chromosome1.length(), _chromosome2.length()); ++i)
 		count += static_cast<size_t>(_chromosome1.optimisationSteps()[i] != _chromosome2.optimisationSteps()[i]);
 
-	return count + abs(static_cast<int>(_chromosome1.length() - _chromosome2.length()));
+	return count + static_cast<size_t>(abs(
+		static_cast<long>(_chromosome1.length()) -
+		static_cast<long>(_chromosome2.length())
+	));
 }
 
 TemporaryDirectory::TemporaryDirectory(std::string const& _prefix):
